+++ app/types/sorteador.ts
@@ -1,0 +1,11 @@
+export interface dadosSorteadorNumeros {
+    min: number;
+    max:number;
+    quantidade: number;
+    permitirRepeticao: boolean;
+}
+
+export interface dadosSorteadorNome {
+    nomes: string[];
+    quantidadeGrupos: number;
+}